fix a few warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@20602 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrgeometry.cpp
@@ -1564,7 +1564,7 @@ const char * OGRToOGCGeomType( OGRwkbGeometryType eGeomType )
 const char *OGRGeometryTypeToName( OGRwkbGeometryType eType )
 
 {
-    switch( eType )
+    switch( (int)eType )
     {
       case wkbUnknown:
         return "Unknown (any)";
